requirements.txt
pre-commit==4.2.*
ipywidgets==8.1.*
coqui-tts==0.27.*
f5-tts==1.1.*
gradio==5.*
